% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{create_dep_file}
\alias{create_dep_file}
\title{Create dependencies file}
\usage{
create_dep_file(root_dir, file_name = "Dependencies")
}
\arguments{
\item{root_dir}{\code{character[1]}\cr The root directory}

\item{file_name}{\code{character[1]}\cr File path (relative to the root
directory), default: \code{Dependencies}}
}
\description{
This function collects dependency information for all scripts below the
root directory, and writes a \code{make}-compatible dependencies file.
}

