% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{create_makefile}
\alias{create_makefile}
\title{Create Makefile}
\usage{
create_makefile(root_dir, file_name = "Makefile",
  dep_file_name = "Dependencies")
}
\arguments{
\item{root_dir}{\code{character[1]}\cr The root directory}

\item{file_name}{\code{character[1]}\cr File path (relative to the root
directory), default: \code{Makefile}}

\item{dep_file_name}{\code{character[1]}\cr File path (relative to the root
directory), default: \code{Dependencies}}
}
\description{
This function creates a Makefile that builds all scripts below the root
directory.
}

