% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{create_script}
\alias{create_script}
\title{Creates a script that will run all scripts in the network}
\usage{
create_script(root_dir, filename = "darn.R")
}
\arguments{
\item{root_dir}{\code{character[1]}\cr The root directory}

\item{filename}{\code{[character(1)]}\cr
The file name of the script to be created}
}
\description{
The scripts will be run in an order that guarantees that all prerequisites
are available when needed.
}

