% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{done}
\alias{done}
\title{Write the results of a stage}
\usage{
done(..., .compress = FALSE)
}
\arguments{
\item{...}{\code{[any]}\cr
Objects to save}

\item{.compress}{\code{logical(1)}\cr
Compress output (default: \code{FALSE})}
}
\value{
A named list that contains path information about the current script.
}
\description{
Add a call to this functions as last statement to each of your scripts,
using double-colon notation (\code{darn::done()}).  A script without such a
call is considered to have no output and cannot be a prerequisite of another
script.
}

