% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_target_dir}
\alias{get_target_dir}
\alias{get_target_path}
\title{Returns target directory}
\usage{
get_target_dir()

get_target_path(...)
}
\arguments{
\item{...}{character vectors.}
}
\description{
Can be run by the scripts to determine the write location for additional
files.
}

