% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{init}
\alias{init}
\title{Init a stage}
\usage{
init(..., env_vars = NULL, envir = parent.frame(), path = NULL)
}
\arguments{
\item{...}{[character(1)]\cr
Prerequisites, without path or extension}

\item{env_vars}{\code{[character]}\cr
Environment variables that define the configuration of the script}

\item{envir}{\code{[environment]}\cr
Environment to load the data into. Default: parent frame.}

\item{path}{\code{[character(1)]}\cr
The path of the current script, useful if automatic detection fails.}
}
\value{
A named list that contains path information about the current script.
}
\description{
Add a call to this functions as first statement to each of your scripts,
using double-colon notation (\code{darn::init()}).  A script without such a
call is considered to have no prerequisites.
}

