% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Calls a function, records the first argument}
\usage{
run(file, fun, ...)
}
\arguments{
\item{file}{The name of the R script}

\item{fun}{The function that should be called to process the R script. The
\code{file} argument will be bassed as first argument to that function.}

\item{...}{Additional arguments passed to \code{fun}.}
}
\description{
For use by the \code{Makefile} generated by \code{\link{create_makefile}}.
}
\examples{
\dontrun{
run("A.R", source)
run("A.R", rmarkdown::render)
run("A.R", ezknitr::ezspin)
}
}

