% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{create_makefile}
\alias{create_makefile}
\title{Create Makefile}
\usage{
create_makefile(root_dir, file_name = "Makefile",
  dep_file_name = "Dependencies", src_dir = ".", out_dir = ".",
  env_vars = NULL,
  script = "Rscript -e \\"rmarkdown::render('$<', 'html_document')\\"")
}
\arguments{
\item{root_dir}{\code{[character(1)]}\cr
The root directory}

\item{file_name}{\code{[character(1)]}\cr
File path (relative to the root directory), default: \code{"Makefile"}}

\item{dep_file_name}{\code{[character(1)]}\cr
File path (relative to the root directory), default: \code{"Dependencies"}}

\item{src_dir, out_dir}{\code{[character(1)]}\cr
Source and output directories (relative to the root directory), default:
\code{"."}}

\item{env_vars}{\code{[named list]}\cr
Default values for environment variables that define configurations of the
script, default: none}

\item{script}{\code{[character]}\cr Script that processes the input files,
default: a call to \code{rmarkdown::\link[rmarkdown]{render}}}
}
\description{
This function creates a Makefile that builds all scripts below the root
directory.
}

