% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{done}
\alias{done}
\title{Write the results of a stage}
\arguments{
\item{...}{Objects to save}
}
\description{
Add a call to this functions as last statement to each of your scripts, using
double-colon notation (\code{darg::done()}).  A script without such a call
is considered to have no output and cannot be a dependency of another script.
}

