% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{init}
\alias{init}
\title{Init a stage}
\arguments{
\item{...}{Dependencies}
}
\description{
Add a call to this functions as first statement to each of your scripts,
using double-colon notation (\code{darg::init()}).  A script without such a
call is considered to have no dependencies.
}

