% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{init}
\alias{init}
\alias{init_}
\title{Init a stage}
\usage{
init_(..., .dots = NULL, env_vars = NULL, envir = parent.frame())
}
\arguments{
\item{...}{Dependencies}

\item{.dots}{\code{[list]}\cr
Additional dependencies as list}

\item{env_vars}{\code{[character]}\cr
Environment variables that define the configuration of the script}
}
\description{
Add a call to this functions as first statement to each of your scripts,
using double-colon notation (\code{darn::init()}).  A script without such a
call is considered to have no dependencies.
}

